+++ src/module.js
@@ -15,18 +15,35 @@ export default class Module {
         this._value = null;
     }
 
+    /**
+     * Returns module namespace.
+     * @returns {string} Module namespace.
+     */
     getNamespace() {
         return this._namespace;
     }
 
+    /**
+     * Returns module name.
+     * @returns {string} Module name.
+     */
     getName() {
         return this._name;
     }
 
+    /**
+     * Returns a list of module dependencies.
+     * @returns {(Array<string>|undefined)} List of module dependencies.
+     */
     getDependencies() {
         return this._dependencies;
     }
 
+    /**
+     * Returns initialized module value.
+     * @returns {any} Modules value.
+     * @throws {Error} Throws error if modules is not initialized.
+     */
     getValue() {
         if (!this.isInitialized()) {
             throw new Error('Module is not initialized!');
@@ -39,10 +56,18 @@ export default class Module {
         return this._value;
     }
 
+    /**
+     * Returns value that determines whether module is initialized.
+     * @returns {boolean} Value that determines whether module is initialized.
+     */
     isInitialized() {
         return this._isInitialized;
     }
 
+    /**
+     * Initializes module.
+     * @throws {Error} Throws error if modules is already initialized.
+     */
     initialize(dependencies) {
         if (this.isInitialized()) {
             throw new Error(`Module has been already initialized!`);